feat(api-quality): add reusable helpers and mocks for quality summary tab tests
- Extract shared constants and helper functions to `aq-summary.support.ts`
- Introduce mocks for validation summary API responses (loading, error, and in-progress states)
- Add deterministic test data for rulesets and spec files
- Update documentation to include guidelines for organizing large test specs
