+++ src/tests/portal/00-serial/15-api-quality/api-quality.spec.ts
@@ -1,6 +1,4 @@
-import { registerRulesetFiles, registerVersionFiles, test } from '@fixtures'
-import type { UsedResourcesHelper } from '@fixtures'
-import type { Page } from '@playwright/test'
+import { test } from '@fixtures'
 import {
   type LintRulesetApiType,
   LintRulesetApiTypes,
@@ -8,82 +6,18 @@ import {
   LintRulesetStatuses,
   RULESET_API_TYPE_TITLE_MAP,
   type RulesetWithFile,
-  SERVER_DEFAULT_RULESETS,
   VERSION_API_QUALITY_TAB_REST,
 } from '@portal/entities'
 import { PortalPage } from '@portal/pages'
-import type { RulesetInfoDialog } from '@portal/pages/PortalPage/VersionPage/VersionOverviewTab/OverviewSummaryTab/components/RulesetInfoDialog'
 import { expect, expectFile, expectText } from '@services/expect-decorator'
-import type { LintRulesetsTestDataManager } from '@services/test-data-manager'
 import { formatDateToUI } from '@services/utils'
-import {
-  CLASS_ACTIVE_LINE_NUMBER,
-  CLASS_CODICON_ERROR,
-  CLASS_CODICON_INFO,
-  CLASS_CODICON_WARNING,
-  CLASS_SELECTED_DECORATOR,
-} from '@shared/components/custom/views/RawView'
-import { OPENAPI_ICON, ROOT_RESOURCES, TestFile } from '@shared/entities'
+import { OPENAPI_ICON, type TestFile } from '@shared/entities'
 import { NO_PERM_SEE_PAGE } from '@test-data/portal'
-import { VAR_GR } from '@test-data/portal/groups'
 import { ALIAS_PREFIX } from '@test-data/prefixes'
-import type { Version } from '@test-data/props'
-import { Group, Package } from '@test-data/props'
-import { HOOK_PUBLISH_TIMEOUT } from '@test-setup'
-import path from 'node:path'
-
-/**
- * Options for test resource registration.
- */
-type TestResourcesOptions = {
-  /** Ruleset(s) used in the test */
-  rulesets?: RulesetWithFile | RulesetWithFile[]
-  /** Version(s) whose files should be attached */
-  versions?: Version | Version[]
-}
-
-/**
- * Registers resource files for linter tests.
- * Attaches ruleset and/or version files for debugging on test failure.
- *
- * NOTE: Only register resources when the test outcome depends on them.
- * Tests with mocked APIs (e.g., mocked validation responses) don't need resources
- * since the actual file content doesn't affect the test result.
- */
-const registerTestResources = (
-  usedResources: UsedResourcesHelper,
-  options: TestResourcesOptions,
-): void => {
-  if (options.rulesets) {
-    registerRulesetFiles(usedResources, options.rulesets)
-  }
-
-  if (options.versions) {
-    registerVersionFiles(usedResources, options.versions)
-  }
-}
-
-// Global helper functions
-const activateDefaultRulesetsAndCleanup = async (
-  lintRulesetTdm: LintRulesetsTestDataManager,
-  testIdN: string,
-): Promise<void> => {
-  // Activate default rulesets for all API types used in tests
-  const apiTypes = [LintRulesetApiTypes.OAS_3_0, LintRulesetApiTypes.OAS_3_1]
-  for (const apiType of apiTypes) {
-    const defaultRuleset = await lintRulesetTdm.getRulesetByName({
-      rulesetName: SERVER_DEFAULT_RULESETS[apiType],
-      apiType: apiType,
-    })
-    if (defaultRuleset) {
-      await lintRulesetTdm.activateRuleset({
-        id: defaultRuleset.id,
-        name: defaultRuleset.name,
-      })
-    }
-  }
-  await lintRulesetTdm.deleteTestRulesets(testIdN)
-}
+import * as aqRm from './aq-rm.support'
+import * as aqShared from './aq-shared.support'
+import * as aqSummary from './aq-summary.support'
+import * as aqTab from './aq-tab.support'
 
 test.describe('API Quality Validation', () => {
   const testIdN = process.env.TEST_ID_N!
@@ -91,84 +25,20 @@ test.describe('API Quality Validation', () => {
   // Shared constants
   const { ACTIVE: STATUS_ACTIVE, INACTIVE: STATUS_INACTIVE } = LintRulesetStatuses
   const currentFormattedDate = formatDateToUI(new Date())
-  const OAS_30_LABEL = RULESET_API_TYPE_TITLE_MAP[LintRulesetApiTypes.OAS_3_0]
-  const OAS_31_LABEL = RULESET_API_TYPE_TITLE_MAP[LintRulesetApiTypes.OAS_3_1]
+  const { OAS_30_LABEL, OAS_31_LABEL } = aqShared
 
-  // Shared test data entities
-  const G_AQ = new Group({
-    name: 'API-Quality',
-    alias: 'GAQ',
-    parent: VAR_GR,
-  })
+  // Shared test data entities / resource files
+  const { G_AQ, FILE_GRAPHQL } = aqShared
 
-  // Shared resource files
-  const ROOT_API_QUALITY = path.join(ROOT_RESOURCES, 'portal', 'api-quality')
-  const FILE_GRAPHQL = new TestFile(path.join(ROOT_API_QUALITY, 'specs', 'aq-graphql.graphql'))
-
-  // Shared helper functions - Mocks
-  const mockSystemConfigurationToDisableLinter = async (page: Page): Promise<void> => {
-    await test.step('Mock system configuration API to disable linter', async () => {
-      await page.route('**/api/v2/system/configuration', async (route) => {
-        const response = await route.fetch()
-        const json = await response.json()
-
-        const filteredExtensions = json.extensions.filter(
-          (ext: { name: string }) => ext.name !== 'api-linter',
-        )
-
-        await route.fulfill({
-          status: response.status(),
-          headers: response.headers(),
-          body: JSON.stringify({
-            ...json,
-            extensions: filteredExtensions,
-          }),
-        })
-      })
-    })
-  }
-
-  // Shared helper functions - Assertions
-  const verifyRulesetInfoDialogContent = async (
-    portalPage: PortalPage,
-    ruleset: RulesetWithFile,
-    status: string,
-  ): Promise<void> => {
-    const { rulesetInfoDialog } = portalPage.versionPackagePage
-    const apiTypeLabel = RULESET_API_TYPE_TITLE_MAP[ruleset.apiType]
-
-    await test.step('Verify dialog displays correct content', async () => {
-      await expect(rulesetInfoDialog.title).toContainText(ruleset.name)
-      await expect(rulesetInfoDialog.apiTypeChip).toHaveText(apiTypeLabel)
-      await expect(rulesetInfoDialog.statusChip).toHaveText(status)
-      await expect(rulesetInfoDialog.rulesetFile).toContainText(ruleset.rulesetFile.name)
-    })
-  }
-
-  const verifyRulesetDownload = async (
-    portalPage: PortalPage,
-    ruleset: RulesetWithFile,
-  ): Promise<void> => {
-    const { rulesetInfoDialog } = portalPage.versionPackagePage
-    const downloadedFile = await rulesetInfoDialog.downloadRuleset()
-
-    await test.step('Verify downloaded file has correct content', async () => {
-      await expectFile(downloadedFile).toHaveName(ruleset.rulesetFile.name)
-      await expectFile(downloadedFile).toContainText(ruleset.rulesetFile.testMeta!.yamlString!)
-    })
-  }
-
-  const verifyRulesetCopyLink = async (
-    portalPage: PortalPage,
-    ruleset: RulesetWithFile,
-  ): Promise<void> => {
-    const { rulesetInfoDialog } = portalPage.versionPackagePage
-    const copiedUrl = await rulesetInfoDialog.copyPublicUrl()
-
-    await test.step('Verify clipboard contains URL matching expected pattern', async () => {
-      await expectText(copiedUrl).toContain(`/api-linter/api/v1/rulesets/${ruleset.id}/data`)
-    })
-  }
+  // Shared helper functions
+  const {
+    activateDefaultRulesetsAndCleanup,
+    mockSystemConfigurationToDisableLinter,
+    registerTestResources,
+    verifyRulesetCopyLink,
+    verifyRulesetDownload,
+    verifyRulesetInfoDialogContent,
+  } = aqShared
 
   test.beforeAll(async ({ apihubTDM }) => {
     // Create shared group for all API Quality tests
@@ -180,32 +50,20 @@ test.describe('API Quality Validation', () => {
   })
 
   test.describe('Ruleset Management', () => {
-    // Constants
-    const DEFAULT_API_TYPE_LABEL = RULESET_API_TYPE_TITLE_MAP[LintRulesetApiTypes.OAS_3_0]
-
-    // Test resource files
-    const FILE_SIMPLE_RULESET = new TestFile(
-      `${ROOT_RESOURCES}/portal/api-quality/rulesets/aq-rm-simple-ruleset.yaml`,
-      {
-        yamlString: 'rules:',
-      },
-    )
-    const FILE_INVALID_EXTENSION = new TestFile(
-      `${ROOT_RESOURCES}/portal/api-quality/rulesets/aq-rm-invalid-extension.txt`,
-    )
-
-    // Messages
-    const MSG_RULESET_CREATED_SUCCESS = (rulesetName: string): string => `${rulesetName} ruleset has been created`
-    const MSG_RULESET_DELETED_SUCCESS = (rulesetName: string): string => `${rulesetName} ruleset has been deleted`
-    const MSG_PUBLIC_URL_COPIED_SUCCESS = 'Public URL copied'
-    const MSG_DUPLICATE_NAME = (rulesetName: string, apiType: string): string =>
-      `Ruleset name ${rulesetName} is not unique for API type ${apiType}`
-    const MSG_INVALID_FILE_FORMAT = 'File format must be YAML'
-
-    // Tooltip messages
-    const TIP_CANNOT_DELETE_ACTIVE = 'Cannot delete active ruleset'
-    const TIP_CANNOT_DELETE_WITH_HISTORY =
-      'The ruleset cannot be deleted due to existing versions that have been validated against this ruleset'
+    const {
+      DEFAULT_API_TYPE_LABEL,
+      FILE_INVALID_EXTENSION,
+      FILE_SIMPLE_RULESET,
+      MSG_DUPLICATE_NAME,
+      MSG_INVALID_FILE_FORMAT,
+      MSG_PUBLIC_URL_COPIED_SUCCESS,
+      MSG_RULESET_CREATED_SUCCESS,
+      MSG_RULESET_DELETED_SUCCESS,
+      TIP_CANNOT_DELETE_ACTIVE,
+      TIP_CANNOT_DELETE_WITH_HISTORY,
+      navigateToRulesetManagement,
+      setupRulesetManagementRulesets,
+    } = aqRm
 
     // Ruleset data
     let RUL_INACTIVE_OAS30_N: { id: string; name: string }
@@ -213,54 +71,13 @@ test.describe('API Quality Validation', () => {
     let RUL_INACTIVE_OAS31_N: { id: string; name: string }
     let RUL_GENERAL_OAS30_N: { id: string; name: string }
 
-    // Helper functions - Actions
-    const navigateToRulesetManagement = async (portalPage: PortalPage): Promise<void> => {
-      await test.step('Navigate to Ruleset Management tab', async () => {
-        await portalPage.goto('/portal/settings/rulesets')
-      })
-    }
-
     test.beforeAll(async ({ lintRulesetTdm }) => {
-      const rulesetNamePrefix = `${ALIAS_PREFIX}-`
-
-      const inactiveOas30 = await lintRulesetTdm.createRuleset({
-        rulesetName: `${rulesetNamePrefix}Inactive-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SIMPLE_RULESET,
-      })
-      RUL_INACTIVE_OAS30_N = { id: inactiveOas30.id, name: inactiveOas30.name }
-
-      const inactiveOas31 = await lintRulesetTdm.createRuleset({
-        rulesetName: `${rulesetNamePrefix}Inactive-OAS31-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_1,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SIMPLE_RULESET,
-      })
-      RUL_INACTIVE_OAS31_N = { id: inactiveOas31.id, name: inactiveOas31.name }
-
-      const generalOas30 = await lintRulesetTdm.createRuleset({
-        rulesetName: `${rulesetNamePrefix}General-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SIMPLE_RULESET,
-      })
-      RUL_GENERAL_OAS30_N = { id: generalOas30.id, name: generalOas30.name }
-
-      const previouslyActive = await lintRulesetTdm.createRuleset({
-        rulesetName: `${rulesetNamePrefix}PreviouslyActive-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SIMPLE_RULESET,
-      })
-      RUL_PREVIOUSLY_ACTIVE_OAS30_N = { id: previouslyActive.id, name: previouslyActive.name }
-
-      // Establish activation history: activate PREVIOUSLY_ACTIVE_RULESET_OAS30_N three times
-      // (last activation is shown in table, previous ones in tooltip)
-      for (let i = 0; i < 3; i++) {
-        await lintRulesetTdm.activateRuleset(RUL_PREVIOUSLY_ACTIVE_OAS30_N)
-        await lintRulesetTdm.activateRuleset(RUL_GENERAL_OAS30_N)
-      }
+      ({
+        RUL_INACTIVE_OAS30_N,
+        RUL_PREVIOUSLY_ACTIVE_OAS30_N,
+        RUL_INACTIVE_OAS31_N,
+        RUL_GENERAL_OAS30_N,
+      } = await setupRulesetManagementRulesets(lintRulesetTdm, testIdN))
     })
 
     test.describe('Initial State and Core UI', () => {
@@ -821,238 +638,43 @@ test.describe('API Quality Validation', () => {
   })
 
   test.describe('Quality Summary Tab', () => {
-    // Test resource files
-    const FILE_SUMMARY_RULESET = new TestFile(path.join(ROOT_API_QUALITY, 'rulesets', 'aq-summary-ruleset.yaml'), {
-      yamlString: 'rules:',
-    })
-    const FILE_SIMPLE_RULESET = new TestFile(
-      path.join(ROOT_API_QUALITY, 'rulesets', 'aq-rm-simple-ruleset.yaml'),
-      {
-        yamlString: 'rules:',
-      },
-    )
-    const FILE_SUMMARY_OAS30 = new TestFile(path.join(ROOT_API_QUALITY, 'specs', 'aq-summary-oas30.yaml'))
-    const FILE_SUMMARY_OAS31 = new TestFile(path.join(ROOT_API_QUALITY, 'specs', 'aq-summary-oas31.yaml'))
+    const {
+      FAILED_DOCS_COUNT,
+      ISSUE_COUNTS_CHANGED,
+      ISSUE_COUNTS_INITIAL,
+      MOCK_FAILED_DOC_1,
+      MOCK_FAILED_DOC_2,
+      MSG_CHECKING_VALIDATION,
+      MSG_NO_VALIDATION_RESULTS,
+      MSG_VALIDATION_IN_PROGRESS,
+      MULTI_DOC_ISSUE_COUNTS_CHANGED,
+      MULTI_DOC_ISSUE_COUNTS_INITIAL,
+      PKG_AQ_SUMMARY_N,
+      TIP_ISSUE_ERROR,
+      TIP_ISSUE_HINT,
+      TIP_ISSUE_INFO,
+      TIP_ISSUE_WARNING,
+      TIP_VALIDATION_FAILED,
+      V_MIXED_REST_GQL_N,
+      V_MULTI_SPEC_N,
+      V_OAS30_N,
+      mockValidationSummaryError,
+      mockValidationSummaryInProgress,
+      mockValidationSummaryLoading,
+      mockValidationSummaryNotFound,
+      setupQualitySummaryTabTestData,
+    } = aqSummary
 
-    // Messages
-    const MSG_NO_VALIDATION_RESULTS = 'No validation results.'
-    const MSG_CHECKING_VALIDATION = 'Checking validation status...'
-    const MSG_VALIDATION_IN_PROGRESS = 'Validation is in progress, please wait...'
-
-    // Issue count constants for single document tests
-    const ISSUE_COUNTS_INITIAL = {
-      error: '1',
-      warning: '1',
-      info: '1',
-      hint: '1',
-    } as const
-
-    const ISSUE_COUNTS_CHANGED = {
-      error: '0',
-      warning: '1',
-      info: '0',
-      hint: '0',
-    } as const
-
-    // Issue count constants for multi-document tests
-    const MULTI_DOC_ISSUE_COUNTS_INITIAL = {
-      error: '2',
-      warning: '2',
-      info: '2',
-      hint: '2',
-    } as const
-
-    const MULTI_DOC_ISSUE_COUNTS_CHANGED = {
-      error: '1',
-      warning: '2',
-      info: '1',
-      hint: '1',
-    } as const
-
-    // Tooltip messages
-    const TIP_ISSUE_ERROR =
-      'ErrorA critical violation of the OpenAPI specification that must be fixed. These issues break compliance and may prevent the API from functioning or integrating correctly.'
-    const TIP_ISSUE_WARNING =
-      'WarningA significant deviation from recommended practices that should be addressed. While not invalid, it may lead to misunderstandings or misuse by API consumers.'
-    const TIP_ISSUE_INFO =
-      'InfoA non-blocking suggestion to improve clarity, completeness, or usability. These enhancements help make the API more developer-friendly.'
-    const TIP_ISSUE_HINT =
-      'HintAn optional recommendation for advanced design improvements or optimizations. Helps raise the overall quality, consistency, and maintainability of the API.'
-    const TIP_VALIDATION_FAILED =
-      'Validation failed. Some documents could not be processed during quality validation. See information icon below for details about failed documents.'
-
-    // Mock data for failed documents
-    const MOCK_FAILED_DOC_1 = 'failed-doc-1.yaml'
-    const MOCK_FAILED_DOC_2 = 'failed-doc-2.yaml'
-    const FAILED_DOCS_COUNT = '2'
-
-    // Test data entities
-    const PKG_AQ_SUMMARY_N = new Package({
-      name: 'Quality-Summary',
-      alias: 'PAQSUM',
-      parent: G_AQ,
-    })
-
-    const V_OAS30_N: Version = {
-      pkg: PKG_AQ_SUMMARY_N,
-      version: 'v1-oas30',
-      status: 'draft',
-      files: [{ file: FILE_SUMMARY_OAS30 }],
-    }
-
-    const V_MULTI_SPEC_N: Version = {
-      pkg: PKG_AQ_SUMMARY_N,
-      version: 'v2-multi',
-      status: 'draft',
-      files: [
-        { file: FILE_SUMMARY_OAS30 },
-        { file: FILE_SUMMARY_OAS31 },
-      ],
-    }
-
-    const V_MIXED_REST_GQL_N: Version = {
-      pkg: PKG_AQ_SUMMARY_N,
-      version: 'v3-mixed',
-      status: 'draft',
-      files: [
-        { file: FILE_SUMMARY_OAS30 },
-        { file: FILE_GRAPHQL },
-      ],
-    }
-
-    // Ruleset data
     let RUL_SUMMARY_OAS30_N: RulesetWithFile
     let RUL_SUMMARY_OAS31_N: RulesetWithFile
     let RUL_ALT_OAS30_N: { id: string; name: string }
 
-    // Helper functions - Mocks
-    const mockValidationSummaryError = async (page: Page): Promise<void> => {
-      await test.step('Mock validation summary to return error status with failed documents', async () => {
-        await page.route('**/validation/summary', async (route) => {
-          await route.fulfill({
-            status: 200,
-            contentType: 'application/json',
-            body: JSON.stringify({
-              status: 'error',
-              rulesets: [{
-                id: 'mock-ruleset-id',
-                name: 'Mock Ruleset',
-                apiType: LintRulesetApiTypes.OAS_3_0,
-                status: STATUS_ACTIVE,
-              }],
-              documents: [
-                {
-                  slug: 'doc-1',
-                  documentName: MOCK_FAILED_DOC_1,
-                  apiType: LintRulesetApiTypes.OAS_3_0,
-                  status: 'error',
-                },
-                {
-                  slug: 'doc-2',
-                  documentName: MOCK_FAILED_DOC_2,
-                  apiType: LintRulesetApiTypes.OAS_3_0,
-                  status: 'error',
-                },
-              ],
-            }),
-          })
-        })
-      })
-    }
-
-    const mockValidationSummaryNotFound = async (page: Page): Promise<void> => {
-      await test.step('Mock validation summary to return 404 LintResultNotFound', async () => {
-        await page.route('**/validation/summary', async (route) => {
-          await route.fulfill({
-            status: 404,
-            contentType: 'application/json',
-            body: JSON.stringify({
-              code: 'LintResultNotFound',
-              message: 'Lint result not found',
-            }),
-          })
-        })
-      })
-    }
-
-    const mockValidationSummaryLoading = async (page: Page): Promise<void> => {
-      await test.step('Mock validation summary to delay response indefinitely', async () => {
-        await page.route('**/validation/summary', async () => {
-          // Never fulfill - simulate loading state
-          await new Promise(() => {})
-        })
-      })
-    }
-
-    const mockValidationSummaryInProgress = async (page: Page): Promise<void> => {
-      await test.step('Mock validation summary to return inProgress status', async () => {
-        await page.route('**/validation/summary', async (route) => {
-          await route.fulfill({
-            status: 200,
-            contentType: 'application/json',
-            body: JSON.stringify({
-              status: 'inProgress',
-              rulesets: [],
-            }),
-          })
-        })
-      })
-    }
-
     test.beforeAll(async ({ apihubTDM, lintRulesetTdm }) => {
-      // Extended timeout for API publishing
-      test.setTimeout(HOOK_PUBLISH_TIMEOUT)
-
-      // Create package hierarchy (G_AQ is already created at top level)
-      await apihubTDM.createPackage([PKG_AQ_SUMMARY_N])
-
-      // Create custom rulesets for quality validation
-      const oas30Ruleset = await lintRulesetTdm.createRuleset({
-        rulesetName: `${ALIAS_PREFIX}-Summary-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SUMMARY_RULESET,
-      })
-      RUL_SUMMARY_OAS30_N = {
-        id: oas30Ruleset.id,
-        name: oas30Ruleset.name,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        rulesetFile: FILE_SUMMARY_RULESET,
-      }
-
-      const oas31Ruleset = await lintRulesetTdm.createRuleset({
-        rulesetName: `${ALIAS_PREFIX}-Summary-OAS31-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_1,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SUMMARY_RULESET,
-      })
-      RUL_SUMMARY_OAS31_N = {
-        id: oas31Ruleset.id,
-        name: oas31Ruleset.name,
-        apiType: LintRulesetApiTypes.OAS_3_1,
-        rulesetFile: FILE_SUMMARY_RULESET,
-      }
-
-      const altOas30Ruleset = await lintRulesetTdm.createRuleset({
-        rulesetName: `${ALIAS_PREFIX}-Alt-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_SIMPLE_RULESET,
-      })
-      RUL_ALT_OAS30_N = { id: altOas30Ruleset.id, name: altOas30Ruleset.name }
-
-      // Establish activation history
-      await lintRulesetTdm.activateRuleset(RUL_SUMMARY_OAS30_N)
-      await lintRulesetTdm.activateRuleset(RUL_ALT_OAS30_N)
-
-      // Activate final rulesets for tests
-      await lintRulesetTdm.activateRuleset(RUL_SUMMARY_OAS30_N)
-      await lintRulesetTdm.activateRuleset(RUL_SUMMARY_OAS31_N)
-
-      // Publish versions for quality validation tests
-      await apihubTDM.publishVersion(V_OAS30_N)
-      await apihubTDM.publishVersion(V_MULTI_SPEC_N)
-      await apihubTDM.publishVersion(V_MIXED_REST_GQL_N)
+      ({
+        RUL_SUMMARY_OAS30_N,
+        RUL_SUMMARY_OAS31_N,
+        RUL_ALT_OAS30_N,
+      } = await setupQualitySummaryTabTestData(apihubTDM, lintRulesetTdm, testIdN))
     })
 
     test.describe('UI Visibility and Access Control', () => {
@@ -1531,458 +1153,47 @@ test.describe('API Quality Validation', () => {
   })
 
   test.describe('API Quality Tab', () => {
-    // Test resource files
-    const FILE_QUALITY_TAB_RULESET = new TestFile(path.join(ROOT_API_QUALITY, 'rulesets', 'aq-tab-ruleset.yaml'), {
-      yamlString: 'rules:',
-    })
-    const FILE_TAB_OAS30 = new TestFile(path.join(ROOT_API_QUALITY, 'specs', 'aq-tab-large-oas30.yaml'), {
-      yamlString: 'openapi: 3.0.0',
-      jsonString: '"openapi": "3.0.0"',
-    })
-    const FILE_TAB_OAS31 = new TestFile(path.join(ROOT_API_QUALITY, 'specs', 'aq-tab-combo-oas31.yaml'), {
-      yamlString: 'openapi: 3.1.0',
-    })
-
-    // Test data entities
-    const PKG_AQ_TAB_N = new Package({
-      name: 'Quality-Tab',
-      alias: 'PAQTAB',
-      parent: G_AQ,
-    })
-
-    const V_AQ_TAB_MIXED_N: Version = {
-      pkg: PKG_AQ_TAB_N,
-      version: 'v2-tab-mixed',
-      status: 'draft',
-      files: [{ file: FILE_TAB_OAS30 }, { file: FILE_TAB_OAS31 }],
-    }
-
-    const V_AQ_TAB_MULTI_N: Version = {
-      pkg: PKG_AQ_TAB_N,
-      version: 'v1-tab-multi',
-      status: 'draft',
-      files: [
-        { file: FILE_TAB_OAS30 },
-        { file: FILE_TAB_OAS31 },
-        { file: FILE_GRAPHQL },
-      ],
-    }
+    const {
+      FILE_TAB_OAS30,
+      FILE_TAB_OAS31,
+      V_AQ_TAB_MIXED_N,
+      V_AQ_TAB_MULTI_N,
+    } = aqTab
 
     // Ruleset data
     let RUL_QUALITY_TAB_OAS30_N: RulesetWithFile
     let RUL_QUALITY_TAB_OAS31_N: RulesetWithFile
 
-    // Linter messages (shared across Content and Interactions and Problem Tooltip tests)
-    const MSG_ERROR_1 = 'Synthetic Error 1 Found'
-    const MSG_ERROR_2 = 'Synthetic Error 2 Found'
-    const MSG_WARNING_1 = 'Synthetic Warning 1 Found'
-    const MSG_WARNING_2 = 'Synthetic Warning 2 Found'
-    const MSG_INFO_1 = 'Synthetic Info 1 Found'
-    const MSG_INFO_2 = 'Synthetic Info 2 Found'
-    const MSG_HINT_1 = 'Synthetic Hint 1 Found'
-    const MSG_HINT_2 = 'Synthetic Hint 2 Found'
+    const {
+      ISSUE_TEST_CASES,
+      MSG_ERROR_1,
+      MSG_HINT_1,
+      MSG_INFO_1,
+      MSG_OVERLAP_ERROR_1,
+      MSG_WARNING_1,
+      TEXT_OVERLAP_OPERATION,
+    } = aqTab
 
-    // Rule names from aq-tab-ruleset.yaml (shared across Content and Interactions and Problem Tooltip tests)
-    const RULE_ERROR_1 = 'synth-error-1'
-    const RULE_ERROR_2 = 'synth-error-2'
-    const RULE_WARNING_1 = 'synth-warn-1'
-    const RULE_WARNING_2 = 'synth-warn-2'
-    const RULE_INFO_1 = 'synth-info-1'
-    const RULE_INFO_2 = 'synth-info-2'
-    const RULE_HINT_1 = 'synth-hint-1'
-    const RULE_HINT_2 = 'synth-hint-2'
-
-    // Overlap error messages (for multi-rule scenario)
-    const MSG_OVERLAP_ERROR_1 = 'Synthetic Overlap Error 1 Found'
-    const MSG_OVERLAP_ERROR_2 = 'Synthetic Overlap Error 2 Found'
-    const TEXT_OVERLAP_OPERATION = 'OverlapErr Operation'
-
-    // Issue test cases configuration
-    type IssueTestCase = {
-      linterMessage: string
-      problemText: string
-      yamlLineNumber: number
-      jsonLineNumber: number
-      ruleName: string
-      iconClass: string
-    }
-
-    // Test cases for issue navigation and tooltip tests (shared across Content and Interactions and Problem Tooltip tests)
-    const ISSUE_TEST_CASES: IssueTestCase[] = [
-      {
-        linterMessage: MSG_ERROR_1,
-        problemText: 'Operation with Error1',
-        yamlLineNumber: 9,
-        jsonLineNumber: 11,
-        ruleName: RULE_ERROR_1,
-        iconClass: CLASS_CODICON_ERROR,
-      },
-      {
-        linterMessage: MSG_ERROR_2,
-        problemText: 'Operation with Error2',
-        yamlLineNumber: 177,
-        jsonLineNumber: 275,
-        ruleName: RULE_ERROR_2,
-        iconClass: CLASS_CODICON_ERROR,
-      },
-      {
-        linterMessage: MSG_WARNING_1,
-        problemText: 'Operation with Warn1',
-        yamlLineNumber: 51,
-        jsonLineNumber: 77,
-        ruleName: RULE_WARNING_1,
-        iconClass: CLASS_CODICON_WARNING,
-      },
-      {
-        linterMessage: MSG_WARNING_2,
-        problemText: 'Operation with Warn2',
-        yamlLineNumber: 219,
-        jsonLineNumber: 341,
-        ruleName: RULE_WARNING_2,
-        iconClass: CLASS_CODICON_WARNING,
-      },
-      {
-        linterMessage: MSG_INFO_1,
-        problemText: 'Operation with Info1',
-        yamlLineNumber: 93,
-        jsonLineNumber: 143,
-        ruleName: RULE_INFO_1,
-        iconClass: CLASS_CODICON_INFO,
-      },
-      {
-        linterMessage: MSG_INFO_2,
-        problemText: 'Operation with Info2',
-        yamlLineNumber: 261,
-        jsonLineNumber: 407,
-        ruleName: RULE_INFO_2,
-        iconClass: CLASS_CODICON_INFO,
-      },
-      {
-        linterMessage: MSG_HINT_1,
-        problemText: 'Operation with Hint1',
-        yamlLineNumber: 135,
-        jsonLineNumber: 209,
-        ruleName: RULE_HINT_1,
-        iconClass: CLASS_CODICON_ERROR,
-      },
-      {
-        linterMessage: MSG_HINT_2,
-        problemText: 'Operation with Hint2',
-        yamlLineNumber: 303,
-        jsonLineNumber: 473,
-        ruleName: RULE_HINT_2,
-        iconClass: CLASS_CODICON_ERROR,
-      },
-    ]
-
-    // Format type for YAML/JSON tests
-    type FormatType = 'yaml' | 'json'
-
-    // Helper to get line number based on format
-    const getLineNumber = (
-      testCase: IssueTestCase,
-      format: FormatType,
-    ): number => (format === 'yaml' ? testCase.yamlLineNumber : testCase.jsonLineNumber)
-
-    // Helper functions - Actions
-    const navigateToApiQualityTab = async (
-      portalPage: PortalPage,
-      version: Version,
-    ): Promise<void> => {
-      await test.step('Navigate to the API Quality tab', async () => {
-        await portalPage.gotoVersion(version, VERSION_API_QUALITY_TAB_REST)
-      })
-    }
-
-    const switchToTestDocument = async (
-      portalPage: PortalPage,
-      documentName: string,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-
-      await test.step('Switch to test document', async () => {
-        await apiQualityTab.documentSlt.click()
-        await apiQualityTab.documentSlt.getListItem(documentName).click()
-      })
-    }
-
-    const switchToFormat = async (
-      portalPage: PortalPage,
-      format: FormatType,
-    ): Promise<void> => {
-      const { rawView } = portalPage.versionPackagePage.apiQualityTab
-
-      if (format === 'json') {
-        await test.step('Switch to JSON format', async () => {
-          await rawView.jsonBtn.click()
-          await expect(rawView.jsonBtn).toBePressed()
-        })
-      }
-    }
-
-    // Helper functions - Actions
-    const openProblemPopupViaTooltip = async (
-      portalPage: PortalPage,
-      testCase: IssueTestCase,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-      const { problemTooltip } = rawView
-
-      await test.step('Hover over issue marker to show Tooltip and open Problem Popup', async () => {
-        await apiQualityTab.getProblemRow(testCase.linterMessage).click()
-        await expect(rawView.getTextContent(testCase.problemText)).toBeVisible()
-        await rawView.hoverText(testCase.problemText)
-        await expect(problemTooltip).toBeVisible()
-        await problemTooltip.viewProblemBtn.click()
-      })
-    }
-
-    const openProblemPopupForOverlappingIssue = async (
-      portalPage: PortalPage,
-      linterMessage: string,
-      problemText: string,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-
-      await test.step('Navigate to overlapping issue and open Problem Popup', async () => {
-        await apiQualityTab.getProblemRow(linterMessage).click()
-        await expect(rawView.getTextContent(problemText)).toBeVisible()
-        await rawView.hoverText(problemText)
-        await rawView.problemTooltip.viewProblemBtn.click()
-      })
-    }
-
-    const closeAndVerifyProblemPopup = async (portalPage: PortalPage): Promise<void> => {
-      const { problemPopUp } = portalPage.versionPackagePage.apiQualityTab.rawView
-
-      await test.step('Close Problem Popup and verify it is hidden', async () => {
-        await problemPopUp.closeBtn.click()
-        await expect(problemPopUp).toBeHidden()
-      })
-    }
-
-    const openProblemPopupViaAltF8 = async (
-      portalPage: PortalPage,
-      testCase: IssueTestCase,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-      const { page } = portalPage
-
-      await test.step('Position cursor on issue marker and open Popup via Alt+F8', async () => {
-        await apiQualityTab.getProblemRow(testCase.linterMessage).click()
-        await expect(rawView.getTextContent(testCase.problemText)).toBeVisible()
-        await rawView.clickText(testCase.problemText)
-        await page.keyboard.press('Alt+F8')
-      })
-    }
-
-    // Helper functions - Assertions
-    const verifyProblemPopupContent = async (
-      portalPage: PortalPage,
-      testCase: IssueTestCase,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-      const { problemPopUp } = rawView
-
-      await test.step('Verify Problem Popup appears with correct content', async () => {
-        await expect(problemPopUp).toBeVisible()
-        await expect(problemPopUp.message).toContainText(testCase.linterMessage)
-        await expect(problemPopUp.message).toContainText(LintRulesetLinters.SPECTRAL)
-        await expect(problemPopUp.message).toContainText(testCase.ruleName)
-      })
-
-      await test.step('Verify Popup displays problem type icon', async () => {
-        await expect(problemPopUp.iconContainer).toHaveIconClass(testCase.iconClass)
-      })
-
-      await test.step('Verify Popup shows navigation arrows and Close button', async () => {
-        await expect(problemPopUp.nextProblemBtn).toBeVisible()
-        await expect(problemPopUp.previousProblemBtn).toBeVisible()
-        await expect(problemPopUp.closeBtn).toBeVisible()
-      })
-    }
-
-    const verifyValidationIssuesSorting = async (
-      portalPage: PortalPage,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-
-      // According to test plan: Issues should be sorted by severity order: Error -> Warning -> Info -> Hint
-      // Within the same severity, sorted by document position (line/column) from start to end
-      // Expected order: Error1, Error2, Warning1, Warning2, Info1, Info2, Hint1, Hint2
-      const SORTED_ISSUES_BY_SEVERITY = [
-        MSG_ERROR_1,
-        MSG_ERROR_2,
-        MSG_WARNING_1,
-        MSG_WARNING_2,
-        MSG_INFO_1,
-        MSG_INFO_2,
-        MSG_HINT_1,
-        MSG_HINT_2,
-      ]
-
-      await test.step('Verify the expected number of issues', async () => {
-        await expect(apiQualityTab.getProblemRow()).toHaveCount(SORTED_ISSUES_BY_SEVERITY.length)
-      })
-
-      await test.step('Verify issues are sorted by severity order', async () => {
-        for (let i = 0; i < SORTED_ISSUES_BY_SEVERITY.length; i++) {
-          const row = apiQualityTab.getProblemRow(i + 1)
-          await expect(row.messageCell).toContainText(SORTED_ISSUES_BY_SEVERITY[i])
-        }
-      })
-    }
-
-    const verifyIssueNavigationHighlight = async (
-      portalPage: PortalPage,
-      format: FormatType,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-
-      for (const testCase of ISSUE_TEST_CASES) {
-        await test.step(`Navigate to "${testCase.linterMessage}" and verify highlighting`, async () => {
-          const lineNumber = getLineNumber(testCase, format)
-          const lineNumberContainer = rawView.getLineNumberContainer(lineNumber)
-
-          await test.step('Navigate to issue location and verify line number container is visible', async () => {
-            await apiQualityTab.getProblemRow(testCase.linterMessage).click()
-            await expect(lineNumberContainer).toBeInViewport()
-          })
-
-          await test.step('Verify selected line number is active', async () => {
-            await expect(lineNumberContainer).toHaveClass(CLASS_ACTIVE_LINE_NUMBER)
-          })
-
-          await test.step('Verify selected problem is marked with blue indicator', async () => {
-            await expect(lineNumberContainer.marker).toHaveClass(CLASS_SELECTED_DECORATOR)
-          })
-
-          await test.step('Verify problem text is visible in the viewport', async () => {
-            await expect(rawView.getTextContent(testCase.problemText)).toBeInViewport()
-          })
-        })
-      }
-    }
-
-    const verifyTooltipOnHover = async (
-      portalPage: PortalPage,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-      const { problemTooltip } = rawView
-
-      for (const testCase of ISSUE_TEST_CASES) {
-        await test.step(`Verify Tooltip appears on hover over "${testCase.linterMessage}" issue marker`, async () => {
-          await test.step('Locate an issue marker in the Document Viewer', async () => {
-            await apiQualityTab.getProblemRow(testCase.linterMessage).click()
-            await expect(rawView.getTextContent(testCase.problemText)).toBeVisible()
-          })
-
-          await test.step('Hover over the issue marker text', async () => {
-            // Hint tooltips appear on whitespace before text, so we use hoverHintText
-            if (testCase.linterMessage === MSG_HINT_1 || testCase.linterMessage === MSG_HINT_2) {
-              await rawView.hoverHintText(testCase.problemText)
-            } else {
-              await rawView.hoverText(testCase.problemText)
-            }
-          })
-
-          await test.step('Verify Problem Tooltip appears with issue message, Linter version, Rule name and View Problem button', async () => {
-            await expect(problemTooltip).toBeVisible()
-            await expect(problemTooltip).toContainText(testCase.linterMessage)
-            await expect(problemTooltip).toContainText(LintRulesetLinters.SPECTRAL)
-            await expect(problemTooltip).toContainText(testCase.ruleName)
-            // Hint tooltips don't show "View Problem" button - they only provide informational tooltips
-            if (testCase.linterMessage !== MSG_HINT_1 && testCase.linterMessage !== MSG_HINT_2) {
-              await expect(problemTooltip.viewProblemBtn).toBeVisible()
-            } else {
-              await expect(problemTooltip.viewProblemBtn).toBeHidden()
-            }
-          })
-        })
-      }
-    }
-
-    const verifyTooltipDisappears = async (
-      portalPage: PortalPage,
-      page: Page,
-    ): Promise<void> => {
-      const { apiQualityTab } = portalPage.versionPackagePage
-      const { rawView } = apiQualityTab
-      const { problemTooltip } = rawView
-
-      await test.step('Hover over an issue marker to show the Tooltip', async () => {
-        await apiQualityTab.getProblemRow(MSG_ERROR_1).click()
-        await expect(rawView.getTextContent('Operation with Error1')).toBeVisible()
-        await rawView.hoverText('Operation with Error1')
-        await expect(problemTooltip).toBeVisible()
-      })
-
-      await test.step('Move the cursor away from the issue marker to a neutral area', async () => {
-        await page.mouse.move(0, 0)
-      })
-
-      await test.step('Verify Problem Tooltip disappears', async () => {
-        await expect(problemTooltip).toBeHidden()
-      })
-    }
+    const {
+      closeAndVerifyProblemPopup,
+      navigateToApiQualityTab,
+      openProblemPopupForOverlappingIssue,
+      openProblemPopupViaAltF8,
+      openProblemPopupViaTooltip,
+      switchToFormat,
+      switchToTestDocument,
+      verifyIssueNavigationHighlight,
+      verifyProblemPopupContent,
+      verifyTooltipDisappears,
+      verifyTooltipOnHover,
+      verifyValidationIssuesSorting,
+    } = aqTab
 
     test.beforeAll(async ({ apihubTDM, lintRulesetTdm }) => {
-      // Extended timeout for API publishing
-      test.setTimeout(HOOK_PUBLISH_TIMEOUT)
-
-      // Create package hierarchy (G_AQ is already created at top level)
-      await apihubTDM.createPackage([PKG_AQ_TAB_N])
-
-      // Create custom rulesets for API Quality Tab tests
-      const qualityTabRulesetOas30 = await lintRulesetTdm.createRuleset({
-        rulesetName: `${ALIAS_PREFIX}-Quality-Tab-OAS30-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_QUALITY_TAB_RULESET,
-      })
-      RUL_QUALITY_TAB_OAS30_N = {
-        id: qualityTabRulesetOas30.id,
-        name: qualityTabRulesetOas30.name,
-        apiType: LintRulesetApiTypes.OAS_3_0,
-        rulesetFile: FILE_QUALITY_TAB_RULESET,
-      }
-
-      const qualityTabRulesetOas31 = await lintRulesetTdm.createRuleset({
-        rulesetName: `${ALIAS_PREFIX}-Quality-Tab-OAS31-${testIdN}`,
-        apiType: LintRulesetApiTypes.OAS_3_1,
-        linter: LintRulesetLinters.SPECTRAL,
-        rulesetFile: FILE_QUALITY_TAB_RULESET,
-      })
-      RUL_QUALITY_TAB_OAS31_N = {
-        id: qualityTabRulesetOas31.id,
-        name: qualityTabRulesetOas31.name,
-        apiType: LintRulesetApiTypes.OAS_3_1,
-        rulesetFile: FILE_QUALITY_TAB_RULESET,
-      }
-
-      // Activate rulesets for OAS 3.0 and OAS 3.1
-      await lintRulesetTdm.activateRuleset(RUL_QUALITY_TAB_OAS30_N)
-      await lintRulesetTdm.activateRuleset(RUL_QUALITY_TAB_OAS31_N)
-
-      // Publish versions for API Quality Tab tests
-      await apihubTDM.publishVersion(V_AQ_TAB_MIXED_N)
-      await apihubTDM.publishVersion(V_AQ_TAB_MULTI_N)
-
-      // Wait for validation to complete after publishing
-      await lintRulesetTdm.waitForValidationToComplete({
-        packageId: PKG_AQ_TAB_N.packageId,
-        version: V_AQ_TAB_MIXED_N.version,
-      })
-      await lintRulesetTdm.waitForValidationToComplete({
-        packageId: PKG_AQ_TAB_N.packageId,
-        version: V_AQ_TAB_MULTI_N.version,
-      })
+      ({
+        RUL_QUALITY_TAB_OAS30_N,
+        RUL_QUALITY_TAB_OAS31_N,
+      } = await aqTab.setupApiQualityTabTestData(apihubTDM, lintRulesetTdm, testIdN))
     })
 
     test.describe('UI Visibility and Navigation', () => {
@@ -2179,22 +1390,6 @@ test.describe('API Quality Validation', () => {
     })
 
     test.describe('Ruleset and Dialog Interactions', () => {
-      // Helper functions - Actions
-      const navigateToApiQualityTabAndOpenRulesetDialog = async (
-        portalPage: PortalPage,
-        version: Version,
-      ): Promise<RulesetInfoDialog> => {
-        const { apiQualityTab } = portalPage.versionPackagePage
-
-        await navigateToApiQualityTab(portalPage, version)
-
-        await test.step('Open ruleset dialog', async () => {
-          await apiQualityTab.nameLink.click()
-        })
-
-        return portalPage.versionPackagePage.rulesetInfoDialog
-      }
-
       test.beforeEach(async ({ usedResources }) => {
         registerTestResources(usedResources, {
           rulesets: RUL_QUALITY_TAB_OAS31_N,
@@ -2206,10 +1401,7 @@ test.describe('API Quality Validation', () => {
         tag: '@smoke',
       }, async ({ sysadminPage: page }) => {
         const portalPage = new PortalPage(page)
-        const rulesetInfoDialog = await navigateToApiQualityTabAndOpenRulesetDialog(
-          portalPage,
-          V_AQ_TAB_MIXED_N,
-        )
+        const rulesetInfoDialog = await aqTab.navigateToApiQualityTabAndOpenRulesetDialog(portalPage, V_AQ_TAB_MIXED_N)
 
         await verifyRulesetInfoDialogContent(portalPage, RUL_QUALITY_TAB_OAS31_N, STATUS_ACTIVE)
 
@@ -2223,7 +1415,7 @@ test.describe('API Quality Validation', () => {
       test('P-AQ-TAB-RULE-2 Verify Ruleset Dialog Download', async ({ sysadminPage: page }) => {
         const portalPage = new PortalPage(page)
 
-        await navigateToApiQualityTabAndOpenRulesetDialog(portalPage, V_AQ_TAB_MIXED_N)
+        await aqTab.navigateToApiQualityTabAndOpenRulesetDialog(portalPage, V_AQ_TAB_MIXED_N)
 
         await verifyRulesetDownload(portalPage, RUL_QUALITY_TAB_OAS31_N)
       })
@@ -2231,18 +1423,14 @@ test.describe('API Quality Validation', () => {
       test('P-AQ-TAB-RULE-3 Verify Ruleset Dialog Copy Link', async ({ sysadminPage: page }) => {
         const portalPage = new PortalPage(page)
 
-        await navigateToApiQualityTabAndOpenRulesetDialog(portalPage, V_AQ_TAB_MIXED_N)
+        await aqTab.navigateToApiQualityTabAndOpenRulesetDialog(portalPage, V_AQ_TAB_MIXED_N)
 
         await verifyRulesetCopyLink(portalPage, RUL_QUALITY_TAB_OAS31_N)
       })
     })
 
     test.describe('Content and Interactions', () => {
-      // Severity icon IDs
-      const ERROR_ICON = 'ErrorIcon'
-      const WARNING_ICON = 'WarningIcon'
-      const INFO_ICON = 'InfoIcon'
-      const HINT_ICON = 'HintIcon'
+      const { ERROR_ICON, WARNING_ICON, INFO_ICON, HINT_ICON } = aqTab
 
       test.beforeEach(async ({ usedResources }) => {
         registerTestResources(usedResources, {
@@ -2361,60 +1549,7 @@ test.describe('API Quality Validation', () => {
     })
 
     test.describe('Problem Tooltip - Hover Behavior', () => {
-      // Helper function for combo tooltip test
-      const verifyComboTooltip = async (
-        portalPage: PortalPage,
-      ): Promise<void> => {
-        const { apiQualityTab } = portalPage.versionPackagePage
-        const { rawView } = apiQualityTab
-        const { problemTooltip } = rawView
-        const COMBO_TEXT = 'Combo Error1 Warn1 Info1 Hint1'
-
-        await test.step('Locate a marker with multiple issues of different types in the Document Viewer', async () => {
-          await apiQualityTab.getProblemRow(MSG_ERROR_1).click()
-          await expect(rawView.getTextContent(COMBO_TEXT)).toBeVisible()
-        })
-
-        await test.step('Hover over the issue marker', async () => {
-          // Use hoverHintText instead of hoverText because Hint tooltips appear on whitespace before text.
-          // This ensures we can capture all issues including Hint, which would not be visible if hovering directly on the text.
-          await rawView.hoverHintText(COMBO_TEXT)
-        })
-
-        await test.step('Verify Problem Tooltip appears with all issues for that location', async () => {
-          await expect(problemTooltip).toBeVisible()
-          await expect(problemTooltip).toContainText(MSG_ERROR_1)
-          await expect(problemTooltip).toContainText(MSG_WARNING_1)
-          await expect(problemTooltip).toContainText(MSG_INFO_1)
-          await expect(problemTooltip).toContainText(MSG_HINT_1)
-          await expect(problemTooltip.viewProblemBtn).toBeVisible()
-        })
-      }
-
-      // Helper function for multi-rule same type tooltip test
-      const verifyMultiRuleTooltip = async (
-        portalPage: PortalPage,
-      ): Promise<void> => {
-        const { apiQualityTab } = portalPage.versionPackagePage
-        const { rawView } = apiQualityTab
-        const { problemTooltip } = rawView
-
-        await test.step('Locate a marker where one element triggers multiple rules of the same type', async () => {
-          await apiQualityTab.getProblemRow(MSG_OVERLAP_ERROR_1).click()
-          await expect(rawView.getTextContent(TEXT_OVERLAP_OPERATION)).toBeVisible()
-        })
-
-        await test.step('Hover over the issue marker', async () => {
-          await rawView.hoverText(TEXT_OVERLAP_OPERATION)
-        })
-
-        await test.step('Verify Problem Tooltip appears with both error messages from different rules', async () => {
-          await expect(problemTooltip).toBeVisible()
-          await expect(problemTooltip).toContainText(MSG_OVERLAP_ERROR_1)
-          await expect(problemTooltip).toContainText(MSG_OVERLAP_ERROR_2)
-          await expect(problemTooltip.viewProblemBtn).toBeVisible()
-        })
-      }
+      const { verifyComboTooltip, verifyMultiRuleTooltip } = aqTab
 
       test('P-AQ-TAB-TIP-1-YAML Verify Tooltip appears on hover over issue marker in YAML', {
         tag: '@smoke',
@@ -2675,103 +1810,12 @@ test.describe('API Quality Validation', () => {
     })
 
     test.describe('Problem Navigation via Popup', () => {
-      // Helper function to verify navigation between problems
-      const verifyProblemNavigation = async (
-        portalPage: PortalPage,
-        initialTestCase: IssueTestCase,
-        nextTestCase: IssueTestCase,
-      ): Promise<void> => {
-        const { problemPopUp } = portalPage.versionPackagePage.apiQualityTab.rawView
-
-        await test.step('Verify initial problem is displayed', async () => {
-          await expect(problemPopUp.message).toContainText(initialTestCase.linterMessage)
-        })
-
-        await test.step('Navigate to next problem and verify Popup updates', async () => {
-          await problemPopUp.nextProblemBtn.click()
-          await expect(problemPopUp.message).toContainText(nextTestCase.linterMessage)
-          await expect(problemPopUp.iconContainer).toHaveIconClass(nextTestCase.iconClass)
-        })
-
-        await test.step('Navigate to previous problem and verify Popup updates', async () => {
-          await problemPopUp.previousProblemBtn.click()
-          await expect(problemPopUp.message).toContainText(initialTestCase.linterMessage)
-          await expect(problemPopUp.iconContainer).toHaveIconClass(initialTestCase.iconClass)
-        })
-      }
-
-      const verifyKeyboardNavigation = async (
-        portalPage: PortalPage,
-        initialTestCase: IssueTestCase,
-        nextTestCase: IssueTestCase,
-      ): Promise<void> => {
-        const { problemPopUp } = portalPage.versionPackagePage.apiQualityTab.rawView
-        const { page } = portalPage
-
-        await test.step('Verify initial problem is displayed', async () => {
-          await expect(problemPopUp.message).toContainText(initialTestCase.linterMessage)
-        })
-
-        await test.step('Navigate to next problem via F8 and verify Popup updates', async () => {
-          await page.keyboard.press('F8')
-          await expect(problemPopUp.message).toContainText(nextTestCase.linterMessage)
-        })
-
-        await test.step('Navigate to previous problem via Shift+F8 and verify Popup updates', async () => {
-          await page.keyboard.press('Shift+F8')
-          await expect(problemPopUp.message).toContainText(initialTestCase.linterMessage)
-        })
-      }
-
-      const verifySeverityOrderNavigation = async (
-        portalPage: PortalPage,
-      ): Promise<void> => {
-        const { problemPopUp } = portalPage.versionPackagePage.apiQualityTab.rawView
-
-        // Expected navigation order: Error1, Error2, Warning1, Warning2, Info1, Info2, Hint1, Hint2
-        const SORTED_BY_SEVERITY = [
-          MSG_ERROR_1,
-          MSG_ERROR_2,
-          MSG_WARNING_1,
-          MSG_WARNING_2,
-          MSG_INFO_1,
-          MSG_INFO_2,
-          MSG_HINT_1,
-          MSG_HINT_2,
-        ]
-
-        await test.step('Verify initial problem is first Error', async () => {
-          await expect(problemPopUp.message).toContainText(SORTED_BY_SEVERITY[0])
-        })
-
-        for (let i = 1; i < SORTED_BY_SEVERITY.length; i++) {
-          await test.step(`Navigate to problem ${i + 1} and verify severity order`, async () => {
-            await problemPopUp.nextProblemBtn.click()
-            await expect(problemPopUp.message).toContainText(SORTED_BY_SEVERITY[i])
-          })
-        }
-      }
-
-      const verifyOverlappingIssuesNavigation = async (
-        portalPage: PortalPage,
-      ): Promise<void> => {
-        const { problemPopUp } = portalPage.versionPackagePage.apiQualityTab.rawView
-
-        await test.step('Verify Popup shows one of the overlapping errors', async () => {
-          await expect(problemPopUp.message).toContainText(MSG_OVERLAP_ERROR_1)
-        })
-
-        await test.step('Navigate to next problem and verify it follows severity order', async () => {
-          await problemPopUp.nextProblemBtn.click()
-          // Next should be the second overlap error (same severity)
-          await expect(problemPopUp.message).toContainText(MSG_OVERLAP_ERROR_2)
-        })
-
-        await test.step('Navigate back to verify we can return to first overlap error', async () => {
-          await problemPopUp.previousProblemBtn.click()
-          await expect(problemPopUp.message).toContainText(MSG_OVERLAP_ERROR_1)
-        })
-      }
+      const {
+        verifyKeyboardNavigation,
+        verifyOverlappingIssuesNavigation,
+        verifyProblemNavigation,
+        verifySeverityOrderNavigation,
+      } = aqTab
 
       test('P-AQ-TAB-NAV-1-YAML Verify Next and Previous Problem navigation via buttons in YAML', {
         tag: '@smoke',
@@ -2927,25 +1971,7 @@ test.describe('API Quality Validation', () => {
     })
 
     test.describe('Special States', () => {
-      // Message for No Validation Results state (note: no newline/space between sentences in rendered text)
-      const MSG_NO_VALIDATION_RESULTS =
-        'API Quality results are not availablePlease check the Summary tab for validation status'
-
-      // Mock for validation summary to simulate no validation results
-      const mockValidationSummaryNotAvailable = async (page: Page): Promise<void> => {
-        await test.step('Mock validation summary to return 404 (no validation results)', async () => {
-          await page.route('**/validation/summary', async (route) => {
-            await route.fulfill({
-              status: 404,
-              contentType: 'application/json',
-              body: JSON.stringify({
-                code: 'LintResultNotFound',
-                message: 'Lint result not found',
-              }),
-            })
-          })
-        })
-      }
+      const { MSG_TAB_NO_VALIDATION_RESULTS: MSG_NO_VALIDATION_RESULTS, mockValidationSummaryNotAvailable } = aqTab
 
       test('P-AQ-TAB-EDGE-1-M Verify No Validation Results state', {
         tag: '@smoke',
